test: 스토리북 theming 적용
